requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-node==4.0.0
django-react==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
